Skip boolean return values if the function throws GErrors
A boolean return value in a function throwing a GError is conventionally
a success value, which is set to TRUE iff the GError has not been
thrown. We can thus omit it without loss of information, this makes the
generated API nicer.

See https://bugzilla.gnome.org/show_bug.cgi?id=649657 for some
discussion, and the rationale for the (skip) attribute on return values,
which we also support now.