Use freeHaskellFunPtr directly from callbacks
It seems to be safe to use freeHaskellFunPtr directly inside the Haskell
code being freed, see the comment in rts/Adjustor.c in the ghc source:
"
Now here's something obscure for you:

When generating an adjustor thunk that uses the C calling
convention, we have to make sure that the thunk kicks off
the process of jumping into Haskell with a tail jump. Why?
Because as a result of jumping in into Haskell we may end
up freeing the very adjustor thunk we came from using
freeHaskellFunctionPtr(). Hence, we better not return to
the adjustor code on our way out, since it could by then
point to junk.
The fix is readily at hand, just include the opcodes
for the C stack fixup code that we need to perform when
returning in some static piece of memory and arrange
to return to it before tail jumping from the adjustor thunk.
"