Remove some "unsafe" import keywords
Some of these may potentially be reentrant into Haskell code (the
easiest example to hit is probably g_object_set_property, which can
easily trigger a callback), so the import should not be marked "unsafe".

We are playing it very safe here, in particular we are not marking
things like g_malloc0 as "unsafe", since somebody may need to write a
custom memory allocator in Haskell, and the memory allocators in GLib
are pluggable. Better safe than sorry, even if there is probably nobody
crazy enough to try this...