Autodetect module dependencies
The dependencies of the module given in the command line are now
autodetected, and they are autogenerated too.

The main use of this is in order to generate the "_property" accesors.
Before it was problematic to use accessors from two different modules,
since the relevant typeclasses may have overlapping names. Now all
possible property names are collected, and generated at the end,
avoiding this problem.