Deal better with unknown length arrays in methods
Sometimes C array types are not properly annotated, and we cannot
determine their length.

In the case of method arguments, when we fixed the array length location
in the argument list, we did not take this into account, and generated
wrong code. This is fixed now: we take into account unknown length
arguments, and give a proper error when trying to generate the code.