Preliminary support for accessing structure fields
Now it is possible (in many cases, but in particular not yet callback
fields) to read from Haskell the fields of C structures, which are
marshaled into the appropriate objects via helpers.

For instance, for the Gdk.EventButton structure one can call
Gdk.eventButtonReadX to access the "x" field in the structure.