Continue code generation after recoverable errors
It can happen that code generation for a given method (or signal,
callback, etc.) fails, but the failure is local and relatively
isolated. So far we were completely stopping code generation when this
happened. This is somewhat annoying since fixing this required modifying
haskell-gi to update the blacklist, recompiling, and trying again.

In most cases we can simply omit the failing method from being
generated, and move on to generating bindings for the next one, so now
we do that.