Make castTo total
When trying to cast a GObject into a different object the conversion may
fail if the object we are dealing with is not of the right type. Instead
of giving an error when this fails, just make the whole thing return a
Maybe type, with failure encoded in Nothing, and let the caller deal
with this.

In order not to lose informative error handling, add a gtypeName method,
so the caller can print out something informative when things go wrong.