Add support for external config files
It is often the case that some symbols in the introspection data should
be omitted, due to errors in the data, or because they require
features that we do not support (such as caller allocates).

Add the possibility of specifying this in an extra config file passed to
the binding generator. Also remove the built-in ignored methods, since
they can now be better expressed in terms of these config files.