Improve overrides implementation
The filtering of introspection data is now all done at a single
spot (just after loading the introspection data) instead of being
scattered all around.

The ability for renaming symbols has been dropped along the way, since
it was unclear what would it mean exactly, and it was not necessary so
far.