Fix negative enum handling
We used to read enum values as Word64, but they can be signed, fix
this. Note that the implementation still assumes that all the enum
values can be contained in an Int. This is not easy to fix given the
signature of the Enum typeclass, but maybe we could check and warn if
it is not the case (the assumption seems to be correct on linux 64 bit).

Also use the NegativeLiterals extension, so large negative numbers are
properly interpreted.