Improve handling of optional and nullable args
Since recent versions of gobject-introspection there is a difference
between (nullable) arguments, i.e. those that admit NULL as a value,
mapping to Maybe in Haskell, and (optional) out or inout arguments,
which are a pointer to a value (which may be nullable or not) such that
the function being called can deal with the pointer being NULL. This
typically signals that the out or inout argument is optional.

In haskell-gi we always pass a non-NULL pointer for optional arguments,
so the (optional) annotation is not important.

Here we check that the annotations (nullable) and (optional) are not
mixed up, by checking that (nullable) arguments are actually a pointer
in the C side.

Also, deal properly with (nullable) (inout) arguments. gtk_init is one
example of such argument type.