Use a Word* for the signal connector return type
In the bindings of functions involving the id of a signal handler the
explicit size is used (Word32 or Word64), but we were using CULong on
the hand-written routines (on, after, ...). These types should always
agree, but passing signal handlers around required a fromIntegral. We
now use Word* also for the hand-written routines in order to avoid this.