Rework inheritance mechanism
We can give a much nicer mechanism for resolving inheritance using
closed type families. They allow us to recursively "look up" the list
of parent types just by knowing the direct parent of a type.

In practice, this means that instead of having to list all the ancestors
of a type when generating code, listing the direct ancestor is
sufficient. As a nice side bonus, the error messages when a required
ancestor is not present are now more readable.