Rework overloaded attributes
This is a redesign of the internals of overloaded attributes (_label,
_active, etc.) using closed type families and the reworked inheritance
mechanism of the previous commit.

The generated code is smaller and faster to compile, and the type
checker errors in case one tries to perform a forbidden operation on an
attribute (such as writing to a read-only property, for instance) are
easier to understand.

Along the way I removed support for attributes of the form
"widgetSensitive", and now only "_sensitive" exists, since in the new
design they would have been equivalent. The non-overloaded version
"setWidgetSensitive" still exists, and it should be easy to add back the
"widgetSensitive" alias if necessary.