Fix popup menu test
Build the menu once and then reuse it every time we need to pop it
up. In addition to being more efficient this way, if we create a fresh
menu every time we need to show it, it would get destroyed (and thus
disappear suddenly without user action) whenever the garbage collector
runs, since there are no references left to the menu on the Haskell side.