Precompute the attribute, signal and parents lists
Instead of computing the list of parent types, available attributes and
available signals at compile time, do it at code generation time. This
gives substantial time savings for code using haskell-gi (the compile
time for some non-trivial code can easily be one third of the compile
time before this change).

A smaller change that also saves some compile time (although not as
much, but perhaps another 5-10% in non-trivial code) is that we no
longer check HasAttribute when building attribute lists for "new" and
"set". The error messages are slightly longer this way, in case one
tries to set an attribute that is not available for a given object, but
they are still clear enough.