Exceptions for unexpected NULL returns
A common mistake in the introspection data is not to mark a return
pointer as (nullable). This commit adds some convenience functions for
raising an exception when an unexpected NULL is received from a C
function not marked as having a (nullable) return value.

Also add a convenience callocMem function, analogous to the existing
allocMem.