Tighten generated version bounds
Due to a shortcoming in cabal
( https://github.com/haskell/cabal/issues/3016 ) we cannot rely on it to
do package version resolution properly for Haskell packages binding to
external libraries, since it does not take installed versions of system
packages (available via pkg-config) into account: it simply tries to
install the latest version from cabal unconditionally, which will fail
in most systems due to older versions of the libraries being
present (older than the latest version present in hackage, which may be
very recent).

To work around this, we instead do tight binding of the haskell package
version, but keep the pkgconfig-depends field loose. What this means is
that one should depend on the lowest version of the API that is
supported, with an _equality_ constraint in the cabal file.

For example, using build-depends:
gi-gtk == 0.3.16.*
will work with installed gtk versions 3.16.* and up (3.18, 3.20, ...), but
build-depends: gi-gtk >= 0.3.16
will give problems due to the shortcoming in cabal mentioned above.