Modularize the generated bindings
Make the output of the code generator modular. That is, we
create (roughly) a module per Object/Interface/etc.

Right now there is a big "GI.Package.Types" module containing all type
definitions and instances in order to get around the limitations of
Haskell's module system, in particular the difficulties with recursive
modules. It may be possible to improve the situation with hs-boot files.

Overloading of signals and attributes is broken with this commit, this
will be fixed in a later commit.