Make the module structure recursive
For documentation purposes it is better if we can keep all
definitions (including signals and properties) associated with an
object/interface in a single file. This is somewhat problematic since
these definitions may require definitions in other modules.

The standard solution (and what we were doing previously) is to put all
the types in a single ".Types" file, but this is fairly bad for
documentation purposes when we are binding large libraries such as
gtk. We would like to go to the haddock page for `GI.Gtk.Object.Label`,
and see all the information about the Label, including properties,
signals, etc, but with the approach of using a `Types` file signals and
properties must all live in (the gigantic) `GI.Gtk.Types`. In addition,
it would be best if the links to `Label` in the documentation for other
modules took us for the documentation of `GI.Gtk.Object.Label`, and not
`GI.Gtk.Types`.

In order to achieve these goals we need recursive modules, which are
supported in GHC via .hs-boot files.