Add support for overloaded methods
As an example, with the OverloadedLabels extension in ghc 8.0, instead
of writing

> containerAdd win widget

one can write more simply

> #add win widget

(And similarly for any method of any object, interface, struct or
union). There is also support for ghc >= 7.8, replacing the hash by an
underscore, as usual:

> _add win widget

In this case one needs to pregenerate a 'GI.OverloadedLabels' module
containing the "_label" overloaded labels, running

$ haskell-gi -l $PKG

with $PKG the list of modules to use the labels with (Gtk, WebKit, etc).