Compute the minimum base version for generated code
Code generated by haskell-gi is, for the most part, compatible with
ghc >= 7.8. The exception is that sometimes (very rarely, for the
current set of bindings this only happens for Gst) we generate explicit
bidirectional pattern synonyms, which seems to be well supported in
ghc >= 7.10 only.

This commit makes the autogenerated cabal file for the bindings include
the proper bound on base in case such pattern synonyms are used.