Heuristics for caller-allocates
caller-allocates arguments are arguments that the caller allocates, and
the called function modifies. They are marked as 'out' argumens in the
introspection data, we treat them as 'inout' arguments instead. The
semantics are somewhat tricky: for memory management purposes they
should be treated as "in" arguments, but from the point of view of the
exposed API they should be treated as "inout". Unfortunately we cannot
always just assume that they are purely "out", so in many cases the
generated API is somewhat suboptimal (since the initial values are not
important): for example for g_io_channel_read_chars the size of the
buffer to read is determined by the caller-allocates argument. As a
compromise, we assume that we can allocate anything that is not a
TCArray.