Support for nullable properties and fields
Properties (and struct/union fields) represented by pointers in the C
side can be NULL, so in principle we should wrap the corresponding types
in a Maybe type on the Haskell side, unless they are explicitly marked
as not nullable in the introspection data.

Unfortunately there is currently no such support in
gobject-introspection (
https://bugzilla.gnome.org/show_bug.cgi?id=730479 ). We work around this
by trying to identify the corresponding set/get methods of the object,
and read the nullability from there.