Allow for arbitrary GIR attribute overrides
Support new syntax in the .overrides file:

> set-attr pathSpec attrName newValue

where pathSpec is a specification of a path into an element in the GIR
XML file. The syntax is a set of `nodeSpec` separated by forward
slashes, where in turn a `nodeSpec` is of the form
`nodeName@nodeType`. This selects elements with a "name" attribute with
value `nodeName` and type `nodeType`.

For instance, a XML element of the form <namespace name="Gtk" ...> could
be selected with `Gtk@namespace`, or if specifying the node type is not
important, simply `Gtk`. Alternatively, if the node type is important
but not the name, one can write simply `@nodeType`. A full example is

> set-attr Gtk/Label/set_angle/@parameters/angle transfer-ownership none