Override allocation info for PangoAttribute
This should really be a boxed type, since it has proper copy/free
functions, so treat it in this way in the bindings.

Additionally, it is not really meaningful to allocate a zero-filled
PangoAttribute (one should rather use pango_attr_font_descr_new, or its
siblings), so do not generate helpers to do this.

See https://bugzilla.gnome.org/show_bug.cgi?id=766452 for the proper fix.