Use a Setup.hs helper replacing defaultMain
Previously the helper generated a set of UserHooks, to be run with
defaultMainWithHooks. Now we provide directly a new default main.

This is mostly to work around a bug in stack, where it seems to get
confused in the Cabal versions in stackage and GHC differ, and a library
provides a new way of generating hooks.