Support disowning ManagedPtr
These are like an ordinary ForeignPtr, but when disowned the original
finalizer will not run. In practice this serves to remove the wrapped
foreign memory from the purview of the Haskell GC, even if it was under
our control when constructed. Useful when a destructor is run on a
wrapped ForeignPtr, so we can avoid double-freeing.