Support for disowning managed pointers
This is useful whenever a pointer we manage is passed into a function
which takes ownership of it (for instance
g_dbus_method_invocation_return_value). Previously we were adding a
reference to the object before calling the function taking ownership,
but this is not necessarily safe, since the function may invalidate the
object.