Support for embedded struct/unions
Oftentimes the field inside a struct/union is an embedded struct/union,
instead of a pointer to a different location. Support these properly.

Notice that when reading the field we return a copy of the embedded
struct, so modifications of the returned struct will not affect the
original struct. This is on purpose, in order to increase
safety (otherwise the garbage collector may decide to free the parent
structure while we are modifying the embedded one, and havoc will
ensue).