Make signal connectors type-safe
Add a constraint on the providing type (IsButton for example) for
onTypeSignal and afterTypeSignal connectors (onButtonClicked and
afterButtonClicked for example).

As side effects of these changes, the corresponding connectTypeSignal
connectors are now gone (since they were not particularly useful, and
cluttered the API), but they could be added back easily if
necessary). Also every interface that provides signals or properties is
now assumed to descend from GObject, even if not explicitly marked as
such in the introspection data.