Don't expose GstByte{Reader,Writer} in bindings
They render gi-gstbase unusable since a lot of the functions are written
as static inlines in the headers, meaning the symbols don't exist in
libgstbase-1.0.so.

This should not have a real impact, as we expect that this sort of
functionality is better handled by some other library native to Haskell.