Do not build the haskell-gi executable in haskell-gi.cabal
The `haskell-gi` executable is a thin wrapper over the haskell-gi
library, which is useful for debugging purposes, but not the
recommended way of generating bindings (which is instead to use the
custom Setup.hs code, as the gi-* libraries in hackage do).

So move it to its own cabal file to minimize confusion.