Add a license exception for static linking
Add an exception to the LGPL for the generated libraries to allow static
linking.

Normally, under the terms of the LGPL, when an application is linked against an
LGPL library, the end user must be able to modify and replace the LGPL part of
the application. This is difficult for statically linked Haskell libraries in
GHC, so application authors that use LGPL code may be obliged to supply source
code for their applications.