Inherit override info
For each library that we bind we now generate a GI.PkgName.Config
module, exporting a symbol 'overrides' that is a copy of the overrides
file given when constructing the binding.

We load this module during the configure step to apply the overrides
for the dependencies of the current module.

This is necessary in order to propagate some particularly important
fixes from a given binding to other bindings using it, see for example
https://github.com/haskell-gi/haskell-gi/issues/218