Add an allocating setting operator (:&=)
This can be used as

set object [attrName :&= value]

This operator can be helpful in situations in which the Haskell -> C
conversion requires allocating memory, and the object owning the
attribute cannot take care of memory management. The (:&=) allows for
setting attributes conveniently even if the conversion could in
principle leak memory.

An example where this is quite useful is when setting callback fields
inside class structs. In this case it is fine if the memory used by
the Haskell -> C conversion leaks, since this is only done once per
GObject type.