Initialize out pointers to null
Some functions, such as poppler_page_get_text_layout, accept an out
argument to a pointer to a dynamically allocated array of objects, but
might not initialize this pointer under some circumstances, for
instance if the number of elements in the array would be 0.

This could lead to crashes, such as the one in [issue
289](https://github.com/haskell-gi/haskell-gi/issues/289). What
happened here is that under some circumstances the number of elements
in the resulting array would be 0, and poppler_page_get_text_layout
would leave the pointer untouched in that case. The marshaling into a
Haskell list would succeed, and give the empty list, but when we freed
the array we would crash, as the contents of the memory area that we
reserved for the pointer would be undefined.

We now initialize these memory areas to NULL. freeMem(NULL) does
nothing, so we avoid the crash.