Rework type hierarchy
In Gtk4 GdkEvent and descendants are objects, with an associated type,
but they are not descendants of GObject. Accomodating these properly
requires reworking the internal type hierarchy of haskell-gi, we now
have:

* TypedObject : For types with an associated GType

* ManagedPtrNewtype : Types wrapped inside a ManagedPtr (a thin
  wrapper over ForeignPtr)

* ManagedPtrNewtype a => BoxedPtr a : For memory regions that we know
  how to allocate/deallocate manually.

* (ManagedPtrNewtype a, TypedObject a) => GBoxed a : For types
  registered as boxed types in GLib.

* (ManagedPtrNewtype a, TypedObject a) => GObject a : for types
  descending from GObject.

These changes have a number of knock-on effects, although luckily
without much visible effects outside the bindings. The main one is
that one needs to replace

* gobjectType (in GObject) -> glibType (in TypedObject)

Additionally, GValue is now a fully fundamental type in the bindings.