Add support for ghc 9.0.1
Due to the simplified subsumption changes in ghc 9.0.1
(https://github.com/ghc-proposals/ghc-proposals/blob/master/proposals/0287-simplify-subsumption.rst)
we need to modify the CodeGen monad to expose the error type explicitly (or alternatively, we would need to eta expand everywhere).

This implies that we cannot do

type CodeGen a = forall e. BaseCodeGen e a

f :: CodeGen ()
...

but instead we need to do

f :: BaseCodeGen e ()
...

So do this, and rename BaseCodeGen -> CodeGen, as there is now no
reason for these two types to be different.

This is a change in the exposed API of haskell-gi, so bump the
versions of haskel-gi and haskell-gi-base to 0.25.0.