Support an implicit ?self parameter in signal handlers
In general we omit the "self" parameter in signal handlers in the
Haskell bindings. So if, for instance, the signal is
Gtk.Button::clicked, and the C handler has type

void clicked_cb(GtkButton *self, gpointer user_data)

the Haskell counterpart will have type "IO ()". The user_data and self
parameters are omitted, because they can be captured easily when
defining the handler, so they are fairly redundant.

This has worked well in general, but unfortunately becomes a problem
in Gtk4, where the event handling goes via event controllers. Here you
create an event controller, connect to some signals from it, and then
add the event controller to a widget. The way this last step is
implemented is that the widget takes ownership of the event
controller, so further access from the Haskell side is in principle
unsafe, and haskell-gi will warn about it. See

https://github.com/haskell-gi/haskell-gi/issues/346

for an example where this happens in practice.

Fixing this is not straightforward: we cannot just change the type of
the callbacks on the Haskell side, as this will lead to a lot of
breakage.

One option is to add variants of the connectors (on, after, on*,
after*) that accept an extra argument, but this will add more clutter
to an already busy set of functions for connecting to signals.

Another option is to mark certain signals as requiring access to
"self". This is doable, but it requires maintenance by hand, and a
fairly detailed understanding of the API being wrapped.

What we do here instead is add an implicit parameter ?self to the type
of the callback. This is backwards compatible (no existing callbacks
should need to change), and allows to access the "self" parameter in
any signal callback if needed. So in the example of the event
controller, we can do:

  after controller #pressed $ \nPress x y -> do
    button <- #getCurrentButton ?self
    putStrLn $ "Button pressed: " <> show nPress <> " "
      <> show x <> " " <> show y <> " button: " <> show button

or more explicitly:

pressedCB :: (?self :: Gtk.GestureClick) => Int32 -> Double -> Double -> IO ()
pressedCB nPress x y = do
    button <- #getCurrentButton ?self
    putStrLn $ "Button pressed: " <> show nPress <> " "
      <> show x <> " " <> show y <> " button: " <> show button

... and then after creating the controller ...

  after controller #pressed pressedCB

See examples/Gtk4/gtk-example.hs for a fully explicit example.