Bump GtkSheet version to 4.3.5 (#365)
* Bump GtkSheet version to 4.3.5

This version includes some fixes, including to the GIR files used for
binding generation.

* Fix CI by setting libdir prefix

This is about the same in terms of a solution as the old `mv` was, since
it still requires us to manually intervene to make sure the .typelib
file ends up in the right place. But now that the file doesn't also need
renaming, using `libdir` is slightly more principled.

If you're curious why this is necessary at all, [this
commit](https://gitlab.gnome.org/GNOME/gobject-introspection/-/commit/b9d269d4bf0b85c3e1a6d17136a33ad5a8c228c4)
from 2010 is responsible. GtkSheet relies on that gobject-introspection
Makefile for its own build process, and at some point the Makefile was
changed to not respect the system-level configuration for where .typelib
files should go. Since it instead relies on `libdir`, we can alter the
value of `libdir` to get the files in the right place.