Fix annotations for user_data arguments in callbacks
The last argument of callbacks is often a user_data argument, but
currently gobject-introspection does not have an annotation
representing this. This is generally OK, since the gir generator will
mark these arguments as (closure) if they are named user_data, and we
do the right things in this case, but recently there has been a push
to "fix" these annotations by removing them without providing any
replacement, which breaks the bindings. See

https://github.com/haskell-gi/haskell-gi/issues/405

for a place where this leads to problems, and

https://gitlab.gnome.org/GNOME/gobject-introspection/-/issues/450

for the proposal on how to fix the issue on the gobject-introspection
side.

Until that is fixed properly in the gobject-introspection side, we try
to guess which arguments in callbacks are user_data arguments.