Name version gdk, gtk and others to allow for easier parallel installation
As explained in

https://github.com/haskell-gi/haskell-gi/issues/329

the fact that we are providing bindings for gtk3 and gtk4 in a single
package (gi-gtk-3.0.* and gi-gtk-4.0.*) causes problems for
distributions and stackage, since they typically don't allow for
installing multiple versions of a single package at one time.

This commit is a first step towards allowing for parallel installation
in these settings: we now provide two distinct packages gi-gtk3 and
gi-gtk4, and similarly for other libraries such as gdk, or webkit.