Cast pointer when transferring untyped closures
GClosures of unknown type are represented as 'GClosure a' on the
Haskell side (with 'a' a free type parameter), while they are
represented as 'GClosure ()' on the types of the FFI function.

So a pointer cast was needed when calling the foreign function, but
this was missing in the case in which we transfer ownership of the
closure to the FFI function being called.